check perms in die

+++ src/commands/dev/die.ts
@@ -1,5 +1,6 @@
 import Command from '@core/command';
-import { e } from '@core/vars';
+import { pick } from '@core/utils';
+import { author, e, noPermissionAnswers } from '@core/vars';
 
 export = {
   name: 'die',
@@ -7,6 +8,11 @@ export = {
     brief: 'i wil stop runing'
   },
   async execute() {
+    if (this.author.id !== author) {
+      this.reply(pick(noPermissionAnswers));
+
+      return;
+    }
     await this.reply(`i shal returne! ${e.salute}`);
     await this.bot.client.destroy();
   }
